Renamed the new common folder under ProTx to shared in order to avoid confusion with the _common folder used by the project.
