+++ client/src/components/ProTx/components/charts/MaltreatmentTypesPlot.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Plot from 'react-plotly.js';
-import { getFipsIdName, capitalizeString } from '../common/dataUtils';
-import { getMaltreatmentPlotData } from '../common/plotUtils';
+import { getFipsIdName, capitalizeString } from '../shared/dataUtils';
+import { getMaltreatmentPlotData } from '../shared/plotUtils';
 import DebugPlot from './DebugPlot';
 import './MaltreatmentTypesPlot.css';
