+++ client/src/components/ProTx/components/charts/PredictiveFeaturesPlot.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import Plot from 'react-plotly.js';
-import { getFipsIdName } from '../common/dataUtils';
-import { getPredictiveFeaturesPlotData } from '../common/plotUtils';
+import { getFipsIdName } from '../shared/dataUtils';
+import { getPredictiveFeaturesPlotData } from '../shared/plotUtils';
 
 import DebugPlot from './DebugPlot';
 import './PredictiveFeaturesPlot.css';
