+++ client/src/components/ProTx/components/dashboard/DisplaySelectors.js
@@ -7,7 +7,7 @@ import {
   OBSERVED_FEATURES_TOP_FIELDS,
   SUPPORTED_YEARS
 } from '../data/meta';
-import { compareSimplifiedValueType } from '../common/dataUtils';
+import { compareSimplifiedValueType } from '../shared/dataUtils';
 import './DisplaySelectors.module.scss';
 
 /* Radio buttons for types of values to display in dropdown (see COOKS-110 for next steps) */
