+++ client/src/components/ProTx/components/shared/Variables.css
